backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
  "alembic>=1.18.3",
  "fastapi>=0.111.0",
  "openpyxl>=3.1",
  "bcrypt>=4.0.0",
  "psycopg[binary]>=3.3.2",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.22",
  "uvicorn[standard]>=0.29.0",
  "yt-dlp>=2026.2.21",
]

[project.scripts]
app = "app.main:app"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
